+++ spec/dummy/client/entries/app-kitten.js
@@ -13,6 +13,7 @@ import DropdownButton from 'kitten/components/dropdowns/dropdown-button'
 import Slider from 'kitten/components/form/slider'
 import Select from 'react-select'
 import SelectWithState from 'kitten/components/form/select-with-state'
+import Search from 'kitten/components/form/search'
 
 // Icons
 import GrabberIcon from 'kitten/components/icons/grabber-icon'
@@ -22,6 +23,12 @@ import ExternalRichLink from 'kitten/components/links/external-rich-link'
 
 // Lists
 import NavList from 'kitten/components/lists/nav-list'
+
+// Tours
+import HeaderTour from 'kitten/components/tours/header-tour'
+import Tour from 'kitten/components/tours/tour'
+import TourPopover from 'kitten/components/tours/tour-popover'
+import TourStep from 'kitten/components/tours/tour-step'
 
 // Simulators
 import SimulatorResult from 'kitten/components/simulators/simulator-result'
@@ -46,12 +53,18 @@ ReactOnRails.register({
   DropdownButton,
   ExternalRichLink,
   GrabberIcon,
+  HeaderTour,
   LoanSimulator,
   Slider,
+  Search,
   Select,
   SelectWithState,
   NavList,
   SimulatorResult,
+  SimulatorResult,
+  Tour,
+  TourPopover,
+  TourStep,
   KarlImageExternalRichLink,
   KarlSvgExternalRichLink,
   KarlSimpleDropdown,